face-recognition/requirements.txt
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
click>=8.1.0
python-dotenv>=1.0.0
Flask>=3.0.0
